Fix some arm32 warnings

+++ libr/anal/block.c
@@ -838,34 +838,34 @@ static bool automerge_predecessor_successor_cb(ut64 addr, void *user) {
 		return true;
 	}
 	bool found;
-	RAnalBlock *pred = ht_up_find (ctx->predecessors, (ut64)block, &found);
+	RAnalBlock *pred = ht_up_find (ctx->predecessors, (ut64)(size_t)block, &found);
 	if (found) {
 		if (pred) {
 			// only one predecessor found so far, but we are the second so there are multiple now
-			ht_up_update (ctx->predecessors, (ut64) block, NULL);
+			ht_up_update (ctx->predecessors, (ut64)(size_t) block, NULL);
 		} // else: already found multiple predecessors, nothing to do
 	} else {
 		// no predecessor found yet, this is the only one until now
-		ht_up_insert (ctx->predecessors, (ut64) block, ctx->cur_pred);
+		ht_up_insert (ctx->predecessors, (ut64)(size_t) block, ctx->cur_pred);
 	}
 	return true;
 }
 
 static bool automerge_get_predecessors_cb(void *user, const ut64 k, const void *v) {
 	AutomergeCtx *ctx = user;
-	const RAnalFunction *fcn = (const RAnalFunction *)k;
+	const RAnalFunction *fcn = (const RAnalFunction *)(size_t)k;
 	RListIter *it;
 	RAnalBlock *block;
 	r_list_foreach (fcn->bbs, it, block) {
 		bool already_visited;
-		ht_up_find (ctx->visited_blocks, (ut64)block, &already_visited);
+		ht_up_find (ctx->visited_blocks, (ut64)(size_t)block, &already_visited);
 		if (already_visited) {
 			continue;
 		}
 		ctx->cur_pred = block;
 		ctx->cur_succ_count = 0;
 		r_anal_block_successor_addrs_foreach (block, automerge_predecessor_successor_cb, ctx);
-		ht_up_insert (ctx->visited_blocks, (ut64)block, (void *)ctx->cur_succ_count);
+		ht_up_insert (ctx->visited_blocks, (ut64)(size_t)block, (void *)ctx->cur_succ_count);
 	}
 	return true;
 }
@@ -893,7 +893,7 @@ R_API void r_anal_block_automerge(RList *blocks) {
 		RListIter *fit;
 		RAnalFunction *fcn;
 		r_list_foreach (block->fcns, fit, fcn) {
-			ht_up_insert (relevant_fcns, (ut64)fcn, NULL);
+			ht_up_insert (relevant_fcns, (ut64)(size_t)fcn, NULL);
 		}
 		ht_up_insert (ctx.blocks, block->addr, block);
 	}
@@ -904,11 +904,11 @@ R_API void r_anal_block_automerge(RList *blocks) {
 	// Now finally do the merging
 	RListIter *tmp;
 	r_list_foreach_safe (blocks, it, tmp, block) {
-		RAnalBlock *predecessor = ht_up_find (ctx.predecessors, (ut64)block, NULL);
+		RAnalBlock *predecessor = ht_up_find (ctx.predecessors, (ut64)(size_t)block, NULL);
 		if (!predecessor) {
 			continue;
 		}
-		size_t pred_succs_count = (size_t)ht_up_find (ctx.visited_blocks, (ut64)predecessor, NULL);
+		size_t pred_succs_count = (size_t)ht_up_find (ctx.visited_blocks, (ut64)(size_t)predecessor, NULL);
 		if (pred_succs_count != 1) {
 			// we can only merge this predecessor if it has exactly one successor
 			continue;
@@ -920,7 +920,7 @@ R_API void r_anal_block_automerge(RList *blocks) {
 		RListIter *bit;
 		RAnalBlock *clock;
 		for (bit = it->n; bit && (clock = bit->data, 1); bit = bit->n) {
-			RAnalBlock *fixup_pred = ht_up_find (ctx.predecessors, (ut64)clock, NULL);
+			RAnalBlock *fixup_pred = ht_up_find (ctx.predecessors, (ut64)(size_t)clock, NULL);
 			if (fixup_pred == block) {
 				r_list_push (fixup_candidates, clock);
 			}
@@ -931,10 +931,10 @@ R_API void r_anal_block_automerge(RList *blocks) {
 			// Update number of successors of the predecessor
 			ctx.cur_succ_count = 0;
 			r_anal_block_successor_addrs_foreach (predecessor, count_successors_cb, &ctx);
-			ht_up_update (ctx.visited_blocks, (ut64)predecessor, (void *)ctx.cur_succ_count);
+			ht_up_update (ctx.visited_blocks, (ut64)(size_t)predecessor, (void *)ctx.cur_succ_count);
 			r_list_foreach (fixup_candidates, bit, clock) {
 				// Make sure all previous pointers to block now go to predecessor
-				ht_up_update (ctx.predecessors, (ut64)clock, predecessor);
+				ht_up_update (ctx.predecessors, (ut64)(size_t)clock, predecessor);
 			}
 			// Remove it from the list
 			r_list_split_iter (blocks, it);
